# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Anagram Finder API",
        "description": "API for finding anagrams and managing wordbase",
        "contact": {
            "email": "admin@anagramfinder.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/wordbase/import": {
            "post": {
                "tags": [
                    "Wordbase"
                ],
                "summary": "Import wordbase",
                "description": "Import words from Estonian word list into the database",
                "operationId": "importWordbase",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "force": {
                                        "description": "Force reimport even if wordbase already exists",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wordbase already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Wordbase import completed"
                                        },
                                        "data": {
                                            "properties": {
                                                "words_imported": {
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "timestamp": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Wordbase imported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Wordbase import completed"
                                        },
                                        "data": {
                                            "properties": {
                                                "words_imported": {
                                                    "type": "integer",
                                                    "example": 50000
                                                },
                                                "timestamp": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Wordbase already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Wordbase already exists"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "WORDBASE_EXISTS"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Import failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Import failed"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "IMPORT_FAILED"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}